requirements.txt
# ask-sdk-webservice-support breaks after 1.3.2 https://github.com/alexa/alexa-skills-kit-sdk-for-python/issues/202
ask-sdk-webservice-support==1.3.2

ask_sdk==1.18.0
django-ask-sdk==1.0.0
django-cors-headers==3.13.0
django-ninja==0.20.0
django==4.1.4
icalendar==5.0.4
jdcal==1.4.1
phonetics==1.0.5
python-dateutil==2.8.2
thefuzz[speedup]==0.19.0
uvicorn[standard]==0.20.0
whitenoise==6.3.0
